# optimize array list