```
/////////// before
//3000	    447746 ns/op
//PASS
/////////// after
//10000	    115983 ns/op
//PASS
func BenchmarkArraylistInsert1(b *testing.B){
	maxCount := 1000
	for i := 0; i < b.N; i++ {
		arr := arraylist.New()
		for i:= 0; i < maxCount; i++{
			arr.Insert(i/2, i)
		}
	}
}

/////////// before
//200000	      9194 ns/op
//PASS
/////////// after
//200000	      6830 ns/op
//PASS
func BenchmarkArraylistInsert1(b *testing.B){
	maxCount := 100
	for i := 0; i < b.N; i++ {
		arr := arraylist.New()
		for i:= 0; i < maxCount; i++{
			arr.Insert(i/2, i)
		}
	}
}
```