@emirpasic 2018-06-19
@xtutu which version of golang did you test this with? huge improvements